refactored supportedLocales & locales data
