+++ App.tsx
@@ -5,7 +5,8 @@ import { Provider } from "react-redux";
 import store from "./src/store";
 import locales from "./src/i18n";
 import { i18nContext } from "./src/contexts/i18n";
-import { I18n, SupportedLocales } from "./src/types/context";
+import { I18n } from "./src/types/context";
+import { SupportedLocales } from "./src/types";
 import { getLocale, updateLocale } from "./src/services/storage";
 import { LocaleSelectorScreen } from "./src/screens";
 import Navigator from "./src/navigation/Navigator";
