+++ src/configs/currencies.ts
@@ -1,0 +1,5 @@
+export const currencies = [
+  { key: "EUR", label: "Euro (€)" },
+  { key: "USD", label: "Untied States Dollar ($)" },
+  { key: "EUR", label: "Great British Poud (£)" },
+];
