+++ src/configs/index.ts
@@ -1,0 +1,4 @@
+import { currencies } from "./currencies";
+import { locales } from "./locales";
+
+export { currencies, locales };
