+++ src/i18n/index.ts
@@ -1,10 +1,4 @@
-import en_GB, { name as enGBName } from "./en_GB";
-import pt_PT, { name as ptPTName } from "./pt_PT";
-import { SupportedLocales } from "../types/context";
+import en_GB from "./en_GB";
+import pt_PT from "./pt_PT";
 
 export default { en_GB, pt_PT };
-
-export const locales = [
-  { key: SupportedLocales.en_GB, name: enGBName, flag: ":gb:" },
-  { key: SupportedLocales.pt_PT, name: ptPTName, flag: ":flag-pt:" },
-];
